fixed functionality problem
now have to add all the brands and models to the selected button
