+++ FrontEnd/src/hooks/useOpenModal.js
@@ -1,7 +1,7 @@
 import { useDispatch } from 'react-redux';
+import { open } from '../slices/ModalSlice';
 import { showLogin } from '../slices/LoginRegisterSlice';
 import { showRegister } from '../slices/LoginRegisterSlice';
-import { open } from '../slices/ModalSlice';
 
 function useOpenModal() {
   const dispatch = useDispatch();
@@ -16,15 +16,9 @@ function useOpenModal() {
     dispatch(showRegister());
   }
 
-  function handleSearchButton(handler) {
-    dispatch(open());
-    dispatch(handler());
-  }
-
   return {
     handleLoginButton,
     handleRegisterButton,
-    handleSearchButton,
   };
 }
