+++ FrontEnd/src/utils/OpenModal.jsx
@@ -7,52 +7,8 @@ import Register from '../features/register/RegisterForm';
 import { close } from '../slices/ModalSlice';
 import { hideLogin, hideRegister } from '../slices/LoginRegisterSlice';
 
-import {
-  showCoupe,
-  showBrand,
-  showFuel,
-  showTransmission,
-  showPrice,
-  showYear,
-  showCity,
-  showColor,
-  showHP,
-  showDoors,
-  showExtras,
-  showWheel,
-  hideCoupe,
-  hideBrand,
-  hideFuel,
-  hideTransmission,
-  hidePrice,
-  hideYear,
-  hideCity,
-  hideColor,
-  hideHP,
-  hideDoors,
-  hideExtras,
-  hideWheel,
-} from '../slices/SearchButtonsSlice';
-import SearchBy from '../ui/SearchBy';
-import OpenButtons from './OpenButtons';
-
 function OpenModal() {
   const isOpenModal = useSelector((state) => state.modal.isOpen);
-
-  const {
-    isCoupe,
-    isBrand,
-    isFuel,
-    isTransmission,
-    isPrice,
-    isYear,
-    isCity,
-    isColor,
-    isHP,
-    isDoors,
-    isExtras,
-    isWheel,
-  } = useSelector((state) => state.buttons);
 
   const isLogin = useSelector((state) => state.logReg.isLoginForm);
   const isRegister = useSelector((state) => state.logReg.isRegisterForm);
@@ -64,31 +20,12 @@ function OpenModal() {
     dispatch(hideRegister());
   }
 
-  function handleCloseButtons() {
-    dispatch(close());
-
-    dispatch(hideCoupe());
-    dispatch(hideBrand());
-    dispatch(hideFuel());
-    dispatch(hideTransmission());
-    dispatch(hidePrice());
-    dispatch(hideYear());
-    dispatch(hideCity());
-    dispatch(hideColor());
-    dispatch(hideHP());
-    dispatch(hideDoors());
-    dispatch(hideExtras());
-    dispatch(hideWheel());
-  }
-
   return (
     <div>
       {isOpenModal && (
-        <Modal onClose={() => dispatch(close())}>
+        <Modal onClose={handleCloseLoginRegister}>
           {isLogin && <Login onCloseModal={handleCloseLoginRegister} />}
           {isRegister && <Register onCloseModal={handleCloseLoginRegister} />}
-
-          {isCoupe && <OpenButtons onCloseModal={handleCloseButtons} />}
         </Modal>
       )}
     </div>
